:refactor: moved univeral methods into analysis utils
