+++ .gitignore
@@ -187,3 +187,9 @@ binary_datasets/*
 
 # Ignore conda activation script
 activate_condaenv.sh
+
+# Ignore timings analysis notebook
+scripts/time_analysis.ipynb
+
+# Ignore analysis tables
+tables/*
