+++ scripts/timings.csv
@@ -1,0 +1,25 @@
+,dataset,n_samples,n_metric_samples,min_train_time,max_train_time,min_metric_time,max_metric_time
+3,VG Faucet,2303,500,370.4977,4910.9575,0.021898269653320312,0.23740887641906736
+4,VG Faucet,2303,1000,370.4977,4910.9575,0.033978939056396484,0.34061217308044434
+5,VG Faucet,2303,2303,370.4977,4910.9575,0.054093360900878906,0.7739889621734619
+16,VG Tree,26004,500,4253.413,55628.2145,0.023094892501831055,0.2163538932800293
+17,VG Tree,26004,1000,4253.413,55628.2145,0.02460026741027832,0.4482729434967041
+18,VG Tree,26004,5000,4253.413,55628.2145,0.04244565963745117,0.8238067626953125
+19,VG Tree,26004,10000,4253.413,55628.2145,0.07236480712890625,0.921006441116333
+20,VG Tree,26004,26004,4253.413,55628.2145,0.14277291297912598,1.223006248474121
+24,VG Watch,2825,500,490.2776,6038.8713,0.02557158470153809,0.16657090187072754
+25,VG Watch,2825,1000,490.2776,6038.8713,0.03026747703552246,0.3392188549041748
+26,VG Watch,2825,2825,490.2776,6038.8713,0.043642520904541016,0.7783894538879395
+37,RVL-CDIP,50000,500,9455.519,112982.4341,0.08681249618530273,0.8107156753540039
+38,RVL-CDIP,50000,1000,9455.519,112982.4341,0.0866861343383789,0.8891267776489258
+39,RVL-CDIP,50000,5000,9455.519,112982.4341,0.13448381423950195,2.0480542182922363
+40,RVL-CDIP,50000,10000,9455.519,112982.4341,0.10516357421875,3.185490131378174
+41,RVL-CDIP,50000,50000,9455.519,112982.4341,0.2488243579864502,13.941641569137571
+52,WikiArt,50000,500,8532.4401,109991.7511,0.19261717796325684,4.934730529785156
+53,WikiArt,50000,1000,8532.4401,109991.7511,0.17374777793884277,5.136492729187012
+54,WikiArt,50000,5000,8532.4401,109991.7511,0.22208213806152344,10.783969163894652
+55,WikiArt,50000,10000,8532.4401,109991.7511,0.27793145179748535,18.592451810836792
+56,WikiArt,50000,50000,8532.4401,109991.7511,0.6017515659332275,43.38290023803711
+60,Oxford Pets,4803,500,835.3203,9957.1173,0.14249038696289062,1.6235387325286863
+61,Oxford Pets,4803,1000,835.3203,9957.1173,0.06390833854675293,1.9923715591430664
+62,Oxford Pets,4803,4803,835.3203,9957.1173,0.06417441368103027,3.6920065879821777
