+++ tables/main-combo-regret.tex
@@ -10,31 +10,31 @@ n_pars & \makecell{1.13} & \makecell{1.56} & \makecell{0.46} & \makecell{\textbf
 \hline
 renggli & \makecell{0.23} & \makecell{\textbf{0.03}} & \makecell{4.93} & \makecell{0.12} & \makecell{0.00} & \makecell{0.49} \\
 \hline
-LogME & \makecell{\textbf{0.00}} & \makecell{2.63} & \makecell{8.20} & \makecell{4.28} & \makecell{\textbf{0.00}} & \makecell{0.49} \\
+LogME & \makecell{\textbf{0.00}} & \makecell{3.19} & \makecell{\textbf{0.00}} & \makecell{4.28} & \makecell{\textbf{0.00}} & \makecell{0.49} \\
 \hline
 imagenet-validation & \makecell{1.13} & \makecell{1.56} & \makecell{0.46} & \makecell{\textbf{0.12}} & \makecell{1.62} & \makecell{0.49} \\
 \hline
-n_pars_renggli & \makecell{1.13} & \makecell{1.56} & \makecell{0.46} & \makecell{\textbf{0.12}} & \makecell{1.62} & \makecell{0.49} \\
+n_pars_renggli & \makecell{1.13} & \makecell{\textbf{0.03}} & \makecell{0.46} & \makecell{0.12} & \makecell{1.62} & \makecell{0.49} \\
 \hline
-n_pars_LogME & \makecell{1.13} & \makecell{1.56} & \makecell{0.46} & \makecell{\textbf{0.12}} & \makecell{1.62} & \makecell{0.49} \\
+n_pars_LogME & \makecell{1.13} & \makecell{2.91} & \makecell{5.99} & \makecell{\textbf{0.12}} & \makecell{1.62} & \makecell{0.49} \\
 \hline
 n_pars_imagenet-validation & \makecell{1.13} & \makecell{1.56} & \makecell{0.46} & \makecell{\textbf{0.12}} & \makecell{1.62} & \makecell{0.49} \\
 \hline
-renggli_LogME & \makecell{\textbf{0.00}} & \makecell{2.63} & \makecell{8.20} & \makecell{0.12} & \makecell{\textbf{0.00}} & \makecell{0.49} \\
+renggli_LogME & \makecell{0.45} & \makecell{3.19} & \makecell{\textbf{0.00}} & \makecell{0.12} & \makecell{\textbf{0.00}} & \makecell{0.49} \\
 \hline
-renggli_imagenet-validation & \makecell{1.13} & \makecell{\textbf{0.03}} & \makecell{0.46} & \makecell{0.12} & \makecell{0.00} & \makecell{0.49} \\
+renggli_imagenet-validation & \makecell{\textbf{0.00}} & \makecell{0.03} & \makecell{0.46} & \makecell{0.12} & \makecell{\textbf{0.00}} & \makecell{0.49} \\
 \hline
-LogME_imagenet-validation & \makecell{\textbf{0.00}} & \makecell{2.63} & \makecell{8.20} & \makecell{0.12} & \makecell{1.62} & \makecell{0.49} \\
+LogME_imagenet-validation & \makecell{0.34} & \makecell{0.03} & \makecell{\textbf{0.00}} & \makecell{0.12} & \makecell{\textbf{0.00}} & \makecell{0.49} \\
 \hline
-n_pars_renggli_LogME & \makecell{1.13} & \makecell{1.56} & \makecell{0.46} & \makecell{\textbf{0.12}} & \makecell{1.62} & \makecell{0.49} \\
+n_pars_renggli_LogME & \makecell{1.13} & \makecell{\textbf{0.03}} & \makecell{0.46} & \makecell{0.12} & \makecell{0.00} & \makecell{0.49} \\
 \hline
 n_pars_renggli_imagenet-validation & \makecell{1.13} & \makecell{1.56} & \makecell{0.46} & \makecell{\textbf{0.12}} & \makecell{1.62} & \makecell{0.49} \\
 \hline
 n_pars_LogME_imagenet-validation & \makecell{1.13} & \makecell{1.56} & \makecell{0.46} & \makecell{\textbf{0.12}} & \makecell{1.62} & \makecell{0.49} \\
 \hline
-renggli_LogME_imagenet-validation & \makecell{\textbf{0.00}} & \makecell{2.63} & \makecell{8.20} & \makecell{0.12} & \makecell{\textbf{0.00}} & \makecell{0.49} \\
+renggli_LogME_imagenet-validation & \makecell{\textbf{0.00}} & \makecell{0.03} & \makecell{\textbf{0.00}} & \makecell{0.12} & \makecell{\textbf{0.00}} & \makecell{0.49} \\
 \hline
-n_pars_renggli_LogME_imagenet-validation & \makecell{1.13} & \makecell{1.56} & \makecell{0.46} & \makecell{\textbf{0.12}} & \makecell{1.62} & \makecell{0.49} \\
+n_pars_renggli_LogME_imagenet-validation & \makecell{1.13} & \makecell{\textbf{0.03}} & \makecell{0.46} & \makecell{0.12} & \makecell{1.62} & \makecell{0.49} \\
 \end{tabular}
 \caption*{\\\textit{Largest value for each row presented in bold.}}
 \end{table}
