+++ tables/mean-combo-results.csv
@@ -1,0 +1,16 @@
+,metric_combos,Kendall's Tau,Regret
+0,n_pars,"{'mean': 0.6015577129949433, 'lower': 0.560913459984425, 'upper': 0.6422019660054616}","{'mean': 1.7862244212053382, 'lower': 1.1006964332327278, 'upper': 2.471752409177949}"
+1,renggli,"{'mean': 0.35314123596565095, 'lower': 0.2725542505202177, 'upper': 0.4337282214110842}","{'mean': 1.5954256220363774, 'lower': 0.8189067974833587, 'upper': 2.371944446589396}"
+2,LogME,"{'mean': 0.371973314628808, 'lower': 0.2336345150126794, 'upper': 0.5103121142449366}","{'mean': 2.1591328303535176, 'lower': 1.383099967687889, 'upper': 2.935165693019146}"
+3,imagenet-validation,"{'mean': 0.580210844937813, 'lower': 0.5371639616313926, 'upper': 0.6232577282442335}","{'mean': 1.7862244212053382, 'lower': 1.1006964332327278, 'upper': 2.471752409177949}"
+4,n_pars_renggli,"{'mean': 0.5290753281274923, 'lower': 0.4884866941306405, 'upper': 0.569663962124344}","{'mean': 1.640881711968747, 'lower': 0.9550908372247806, 'upper': 2.3266725867127134}"
+5,n_pars_LogME,"{'mean': 0.5416053257059751, 'lower': 0.49790734701400696, 'upper': 0.5853033043979432}","{'mean': 1.4874615280089183, 'lower': 0.9344006437667768, 'upper': 2.0405224122510597}"
+6,n_pars_imagenet-validation,"{'mean': 0.5528555630497612, 'lower': 0.5111235802129958, 'upper': 0.5945875458865267}","{'mean': 1.7862244212053382, 'lower': 1.1006964332327278, 'upper': 2.471752409177949}"
+7,renggli_LogME,"{'mean': 0.3746153492704123, 'lower': 0.2426400208809969, 'upper': 0.5065906776598277}","{'mean': 1.4989879842965408, 'lower': 0.7238711186430978, 'upper': 2.2741048499499836}"
+8,renggli_imagenet-validation,"{'mean': 0.5213821823475374, 'lower': 0.4732777006986811, 'upper': 0.5694866639963937}","{'mean': 1.2578677762293227, 'lower': 0.5919952988289603, 'upper': 1.923740253629685}"
+9,LogME_imagenet-validation,"{'mean': 0.4559694744347911, 'lower': 0.35695848547406817, 'upper': 0.554980463395514}","{'mean': 0.8278838273402611, 'lower': 0.4552156957056874, 'upper': 1.2005519589748348}"
+10,n_pars_renggli_LogME,"{'mean': 0.5346048513468761, 'lower': 0.49099539848715107, 'upper': 0.5782143042066011}","{'mean': 1.1453213554085242, 'lower': 0.5619505410442834, 'upper': 1.7286921697727649}"
+11,n_pars_renggli_imagenet-validation,"{'mean': 0.5401116497456212, 'lower': 0.49681737338903115, 'upper': 0.5834059261022113}","{'mean': 1.817023980905599, 'lower': 1.1006341297649875, 'upper': 2.5334138320462105}"
+12,n_pars_LogME_imagenet-validation,"{'mean': 0.5543903820740615, 'lower': 0.5166762214806186, 'upper': 0.5921045426675045}","{'mean': 1.370640523031856, 'lower': 0.8819161919156793, 'upper': 1.8593648541480325}"
+13,renggli_LogME_imagenet-validation,"{'mean': 0.461282851291177, 'lower': 0.3685214717452078, 'upper': 0.5540442308371463}","{'mean': 1.0053615277527268, 'lower': 0.4740408651591158, 'upper': 1.5366821903463377}"
+14,n_pars_renggli_LogME_imagenet-validation,"{'mean': 0.5491428837752893, 'lower': 0.5101812888209709, 'upper': 0.5881044787296077}","{'mean': 1.1017992588151646, 'lower': 0.5686791174318115, 'upper': 1.6349194001985177}"
